instalments update

+++ src/main/java/cyou/arfsd/spendbackend/Repositories/ReloadsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import cyou.arfsd.spendbackend.Models.Reloads;
 
-public interface ReloadsRepository extends CrudRepository<Reloads, Long> {
+public interface ReloadsRepository extends CrudRepository<Reloads, Integer> {
     Iterable<Reloads> findByUserid(Integer userId);
 }
